smart-ml-forecaster/requirements.txt
flask
kafka-python
joblib
scikit-learn